Minor fix for signed-ness

+++ util.h
@@ -254,10 +254,10 @@ class Util {
         template <typename T>
         static void reverseWindingOrder(T &arrayData, const MIntArray &faceCounts)
         {
-            int current_index = 0;
-            for(int i = 0; i < (int) faceCounts.length(); i++)
+            unsigned int current_index = 0;
+            for(unsigned int i = 0; i < faceCounts.length(); i++)
             {
-                for(int a = current_index, b = current_index + faceCounts[i] - 1;
+                for(unsigned int a = current_index, b = current_index + faceCounts[i] - 1;
                         a < b; a++, b--)
                 {
                     std::swap(arrayData[a], arrayData[b]);
